Move bag of words est to tests/ directory where it better belongs
