+++ src/model/bag_of_words.py
@@ -37,13 +37,3 @@ def my_bag_of_words(text, words_to_index, dict_size):
         if word in words_to_index:
             result_vector[words_to_index[word]] += 1
     return result_vector
-
-
-def test_my_bag_of_words():
-    words_to_index = {'hi': 0, 'you': 1, 'me': 2, 'are': 3}
-    examples = ['hi how are you']
-    answers = [[1, 1, 0, 1]]
-    for ex, ans in zip(examples, answers):
-        if (my_bag_of_words(ex, words_to_index, 4) != ans).any():
-            return "Wrong answer for the case: '%s'" % ex
-    return 'Basic tests are passed.'
